requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
requests==2.26.0
gunicorn==20.1.0
pyserial==3.5
